+++ src/diffusers/utils/export_utils.py
@@ -3,16 +3,14 @@ import random
 import struct
 import tempfile
 from contextlib import contextmanager
-from typing import List, Optional, Union
+from typing import List, Union
 
 import numpy as np
 import PIL.Image
 import PIL.ImageOps
-import torch
 
 from .import_utils import (
     BACKENDS_MAPPING,
-    is_matplotlib_available,
     is_opencv_available,
 )
 from .logging import get_logger
@@ -140,354 +138,3 @@ def export_to_video(
         img = cv2.cvtColor(video_frames[i], cv2.COLOR_RGB2BGR)
         video_writer.write(img)
     return output_video_path
-
-
-def colormap(
-    image: Union[np.ndarray, torch.Tensor],
-    cmap: str = "Spectral",
-    bytes: bool = False,
-    _force_method: Optional[str] = None,
-) -> Union[np.ndarray, torch.Tensor]:
-    """
-    Converts a monochrome image into an RGB image by applying the specified colormap. This function mimics the behavior
-    of matplotlib.colormaps, but allows the user to use the most discriminative color map "Spectral" without having to
-    install or import matplotlib. For all other cases, the function will attempt to use the native implementation.
-
-    Args:
-        image: 2D tensor of values between 0 and 1, either as np.ndarray or torch.Tensor.
-        cmap: Colormap name.
-        bytes: Whether to return the output as uint8 or floating point image.
-        _force_method:
-            Can be used to specify whether to use the native implementation (`"matplotlib"`), the efficient custom
-            implementation of the "Spectral" color map (`"custom"`), or rely on autodetection (`None`, default).
-
-    Returns:
-        An RGB-colorized tensor corresponding to the input image.
-    """
-    if not (torch.is_tensor(image) or isinstance(image, np.ndarray)):
-        raise ValueError("Argument must be a numpy array or torch tensor.")
-    if _force_method not in (None, "matplotlib", "custom"):
-        raise ValueError("_force_method must be either `None`, `'matplotlib'` or `'custom'`.")
-
-    def method_matplotlib(image, cmap, bytes=False):
-        if is_matplotlib_available():
-            import matplotlib
-        else:
-            return None
-
-        arg_is_pt, device = torch.is_tensor(image), None
-        if arg_is_pt:
-            image, device = image.cpu().numpy(), image.device
-
-        if cmap not in matplotlib.colormaps:
-            raise ValueError(
-                f"Unexpected color map {cmap}; available options are: {', '.join(list(matplotlib.colormaps.keys()))}"
-            )
-
-        cmap = matplotlib.colormaps[cmap]
-        out = cmap(image, bytes=bytes)  # [?,4]
-        out = out[..., :3]  # [?,3]
-
-        if arg_is_pt:
-            out = torch.tensor(out, device=device)
-
-        return out
-
-    def method_custom(image, cmap, bytes=False):
-        arg_is_np = isinstance(image, np.ndarray)
-        if arg_is_np:
-            image = torch.tensor(image)
-        if image.dtype == torch.uint8:
-            image = image.float() / 255
-        else:
-            image = image.float()
-
-        if cmap != "Spectral":
-            raise ValueError("Only 'Spectral' color map is available without installing matplotlib.")
-
-        _Spectral_data = (  # Taken from matplotlib/_cm.py
-            (0.61960784313725492, 0.003921568627450980, 0.25882352941176473),  # 0.0 -> [0]
-            (0.83529411764705885, 0.24313725490196078, 0.30980392156862746),
-            (0.95686274509803926, 0.42745098039215684, 0.2627450980392157),
-            (0.99215686274509807, 0.68235294117647061, 0.38039215686274508),
-            (0.99607843137254903, 0.8784313725490196, 0.54509803921568623),
-            (1.0, 1.0, 0.74901960784313726),
-            (0.90196078431372551, 0.96078431372549022, 0.59607843137254901),
-            (0.6705882352941176, 0.8666666666666667, 0.64313725490196083),
-            (0.4, 0.76078431372549016, 0.6470588235294118),
-            (0.19607843137254902, 0.53333333333333333, 0.74117647058823533),
-            (0.36862745098039218, 0.30980392156862746, 0.63529411764705879),  # 1.0 -> [K-1]
-        )
-
-        cmap = torch.tensor(_Spectral_data, dtype=torch.float, device=image.device)  # [K,3]
-        K = cmap.shape[0]
-
-        pos = image.clamp(min=0, max=1) * (K - 1)
-        left = pos.long()
-        right = (left + 1).clamp(max=K - 1)
-
-        d = (pos - left.float()).unsqueeze(-1)
-        left_colors = cmap[left]
-        right_colors = cmap[right]
-
-        out = (1 - d) * left_colors + d * right_colors
-
-        if bytes:
-            out = (out * 255).to(torch.uint8)
-
-        if arg_is_np:
-            out = out.numpy()
-
-        return out
-
-    if _force_method is None and torch.is_tensor(image) and cmap == "Spectral":
-        return method_custom(image, cmap, bytes)
-
-    out = None
-    if _force_method != "custom":
-        out = method_matplotlib(image, cmap, bytes)
-
-    if _force_method == "matplotlib" and out is None:
-        raise ImportError("Make sure to install matplotlib if you want to use a color map other than 'Spectral'.")
-
-    if out is None:
-        out = method_custom(image, cmap, bytes)
-
-    return out
-
-
-def visualize_depth(
-    depth: Union[
-        PIL.Image.Image,
-        np.ndarray,
-        torch.FloatTensor,
-        List[PIL.Image.Image],
-        List[np.ndarray],
-        List[torch.FloatTensor],
-    ],
-    val_min: float = 0.0,
-    val_max: float = 1.0,
-    color_map: str = "Spectral",
-) -> Union[PIL.Image.Image, List[PIL.Image.Image]]:
-    """
-    Visualizes depth maps, such as predictions of the `MarigoldDepthPipeline`.
-
-    Args:
-        depth (`Union[PIL.Image.Image, np.ndarray, torch.FloatTensor, List[PIL.Image.Image], List[np.ndarray],
-            List[torch.FloatTensor]]`): Depth maps.
-        val_min (`float`, *optional*, defaults to `0.0`): Minimum value of the visualized depth range.
-        val_max (`float`, *optional*, defaults to `0.0`): Maximum value of the visualized depth range.
-        color_map (`str`, *optional*, defaults to `"Spectral"`): Color map used to convert a single-channel
-                  depth prediction into colored representation.
-
-    Returns: `PIL.Image.Image` or `List[PIL.Image.Image]` with depth maps visualization.
-    """
-    if val_max <= val_min:
-        raise ValueError(f"Invalid values range: [{val_min}, {val_max}].")
-
-    def visualize_depth_one(img, idx=None):
-        prefix = "Depth" + (f"[{idx}]" if idx else "")
-        if isinstance(img, PIL.Image.Image):
-            if img.mode != "I;16":
-                raise ValueError(f"{prefix}: invalid PIL mode={img.mode}.")
-            img = np.array(img).astype(np.float32) / (2**16 - 1)
-        if isinstance(img, np.ndarray) or torch.is_tensor(img):
-            if img.ndim != 2:
-                raise ValueError(f"{prefix}: unexpected shape={img.shape}.")
-            if isinstance(img, np.ndarray):
-                img = torch.from_numpy(img)
-            if not torch.is_floating_point(img):
-                raise ValueError(f"{prefix}: unexected dtype={img.dtype}.")
-        else:
-            raise ValueError(f"{prefix}: unexpected type={type(img)}.")
-        if val_min != 0.0 or val_max != 1.0:
-            img = (img - val_min) / (val_max - val_min)
-        img = colormap(img, cmap=color_map, bytes=True)  # [H,W,3]
-        img = PIL.Image.fromarray(img.cpu().numpy())
-        return img
-
-    if depth is None or isinstance(depth, list) and any(o is None for o in depth):
-        raise ValueError("Input depth is `None`")
-    if isinstance(depth, np.ndarray) or torch.is_tensor(depth):
-        if depth.ndim == 2:
-            return visualize_depth_one(depth)
-        if depth.ndim == 3:
-            return [visualize_depth_one(img, idx) for idx, img in enumerate(depth)]
-        if depth.ndim == 4:
-            if depth.shape[1] != 1:
-                raise ValueError(f"Unexpected input shape={depth.shape}, expecting [N,1,H,W].")
-            return [visualize_depth_one(img[0], idx) for idx, img in enumerate(depth)]
-        raise ValueError(f"Unexpected input shape={depth.shape}: expecting 2, 3, or 4 dimensions.")
-    elif isinstance(depth, list):
-        return [visualize_depth_one(img, idx) for idx, img in enumerate(depth)]
-    else:
-        raise ValueError(f"Unexpected input type: {type(depth)}")
-
-
-def export_depth_to_16bit_png(
-    depth: Union[np.ndarray, torch.FloatTensor, List[np.ndarray], List[torch.FloatTensor]],
-    val_min: float = 0.0,
-    val_max: float = 1.0,
-) -> Union[PIL.Image.Image, List[PIL.Image.Image]]:
-    def export_depth_to_16bit_png_one(img, idx=None):
-        prefix = "Depth" + (f"[{idx}]" if idx else "")
-        if not isinstance(img, np.ndarray) and not torch.is_tensor(img):
-            raise ValueError(f"{prefix}: unexpected type={type(img)}.")
-        if img.ndim != 2:
-            raise ValueError(f"{prefix}: unexpected shape={img.shape}.")
-        if torch.is_tensor(img):
-            img = img.cpu().numpy()
-        if not np.issubdtype(img.dtype, np.floating):
-            raise ValueError(f"{prefix}: unexected dtype={img.dtype}.")
-        if val_min != 0.0 or val_max != 1.0:
-            img = (img - val_min) / (val_max - val_min)
-        img = (img * (2**16 - 1)).astype(np.uint16)
-        img = PIL.Image.fromarray(img, mode="I;16")
-        return img
-
-    if depth is None or isinstance(depth, list) and any(o is None for o in depth):
-        raise ValueError("Input depth is `None`")
-    if isinstance(depth, np.ndarray) or torch.is_tensor(depth):
-        if depth.ndim == 2:
-            return export_depth_to_16bit_png_one(depth)
-        if depth.ndim == 3:
-            return [export_depth_to_16bit_png_one(img, idx) for idx, img in enumerate(depth)]
-        if depth.ndim == 4:
-            if depth.shape[1] != 1:
-                raise ValueError(f"Unexpected input shape={depth.shape}, expecting [N,1,H,W].")
-            return [export_depth_to_16bit_png_one(img[0], idx) for idx, img in enumerate(depth)]
-        raise ValueError(f"Unexpected input shape={depth.shape}: expecting 2, 3, or 4 dimensions.")
-    elif isinstance(depth, list):
-        return [export_depth_to_16bit_png_one(img, idx) for idx, img in enumerate(depth)]
-    else:
-        raise ValueError(f"Unexpected input type: {type(depth)}")
-
-
-def visualize_normals(
-    normals: Union[
-        np.ndarray,
-        torch.FloatTensor,
-        List[np.ndarray],
-        List[torch.FloatTensor],
-    ],
-    flip_x: bool = False,
-    flip_y: bool = False,
-    flip_z: bool = False,
-) -> Union[PIL.Image.Image, List[PIL.Image.Image]]:
-    """
-    Visualizes surface normals, such as predictions of the `MarigoldNormalsPipeline`.
-
-    Args:
-        normals (`Union[np.ndarray, torch.FloatTensor, List[np.ndarray], List[torch.FloatTensor]]`): Surface normals.
-        flip_x (`bool`, *optional*, defaults to `False`): Flips the X axis of the normals frame of reference.
-                  Default direction is right.
-        flip_y (`bool`, *optional*, defaults to `False`):  Flips the Y axis of the normals frame of reference.
-                  Default direction is top.
-        flip_z (`bool`, *optional*, defaults to `False`): Flips the Z axis of the normals frame of reference.
-                  Default direction is facing the observer.
-
-    Returns: `PIL.Image.Image` or `List[PIL.Image.Image]` with surface normals visualization.
-    """
-    flip_vec = None
-    if any((flip_x, flip_y, flip_z)):
-        flip_vec = torch.tensor(
-            [
-                (-1) ** flip_x,
-                (-1) ** flip_y,
-                (-1) ** flip_z,
-            ],
-            dtype=torch.float32,
-        )
-
-    def visualize_normals_one(img, idx=None):
-        prefix = "Normals" + (f"[{idx}]" if idx else "")
-        if isinstance(img, np.ndarray) or torch.is_tensor(img):
-            if img.ndim != 3 or img.shape[0] != 3:
-                raise ValueError(f"{prefix}: unexpected shape={img.shape}.")
-            if isinstance(img, np.ndarray):
-                img = torch.from_numpy(img)
-            if not torch.is_floating_point(img):
-                raise ValueError(f"{prefix}: unexected dtype={img.dtype}.")
-        else:
-            raise ValueError(f"{prefix}: unexpected type={type(img)}.")
-        img = img.permute(1, 2, 0)
-        if flip_vec is not None:
-            img *= flip_vec.to(img.device)
-        img = (img + 1.0) * 0.5
-        img = (img * 255).to(dtype=torch.uint8, device="cpu").numpy()
-        img = PIL.Image.fromarray(img)
-        return img
-
-    if normals is None or isinstance(normals, list) and any(o is None for o in normals):
-        raise ValueError("Input normals is `None`")
-    if isinstance(normals, np.ndarray) or torch.is_tensor(normals):
-        if normals.ndim == 3:
-            return visualize_normals_one(normals)
-        if normals.ndim == 4:
-            if normals.shape[1] != 3:
-                raise ValueError(f"Unexpected input shape={normals.shape}, expecting [N,3,H,W].")
-            return [visualize_normals_one(img, idx) for idx, img in enumerate(normals)]
-        raise ValueError(f"Unexpected input shape={normals.shape}: expecting 3 or 4 dimensions.")
-    elif isinstance(normals, list):
-        return [visualize_normals_one(img, idx) for idx, img in enumerate(normals)]
-    else:
-        raise ValueError(f"Unexpected input type: {type(normals)}")
-
-
-def visualize_uncertainty(
-    uncertainty: Union[
-        np.ndarray,
-        torch.FloatTensor,
-        List[np.ndarray],
-        List[torch.FloatTensor],
-    ],
-    saturation_percentile=95,
-) -> Union[PIL.Image.Image, List[PIL.Image.Image]]:
-    """
-    Visualizes dense uncertainties, such as produced by `MarigoldDepthPipeline` or `MarigoldNormalsPipeline`.
-
-    Args:
-        uncertainty (`Union[np.ndarray, torch.FloatTensor, List[np.ndarray], List[torch.FloatTensor]]`):
-            Uncertainty maps.
-        saturation_percentile (`int`, *optional*, defaults to `95`):
-            Specifies the percentile uncertainty value visualized with maximum intensity.
-
-    Returns: `PIL.Image.Image` or `List[PIL.Image.Image]` with uncertainty visualization.
-    """
-
-    def visualize_uncertainty_one(img, idx=None):
-        prefix = "Uncertainty" + (f"[{idx}]" if idx else "")
-        if isinstance(img, np.ndarray) or torch.is_tensor(img):
-            if img.ndim == 2:
-                img = img[None]
-            if img.ndim != 3 or img.shape[0] != 1:
-                raise ValueError(f"{prefix}: unexpected shape={img.shape}.")
-            if isinstance(img, np.ndarray):
-                img = torch.from_numpy(img)
-            if not torch.is_floating_point(img):
-                raise ValueError(f"{prefix}: unexected dtype={img.dtype}.")
-            if img.min() < 0:
-                raise ValueError(f"{prefix}: unexected data range, min={img.min()}.")
-        else:
-            raise ValueError(f"{prefix}: unexpected type={type(img)}.")
-        img = img.squeeze(0).cpu().numpy()
-        saturation_value = np.percentile(img, saturation_percentile)
-        img = np.clip(img * 255 / saturation_value, 0, 255)
-        img = img.astype(np.uint8)
-        img = PIL.Image.fromarray(img)
-        return img
-
-    if uncertainty is None or isinstance(uncertainty, list) and any(o is None for o in uncertainty):
-        raise ValueError("Input uncertainty is `None`")
-    if isinstance(uncertainty, np.ndarray) or torch.is_tensor(uncertainty):
-        if uncertainty.ndim in (2, 3):
-            return visualize_uncertainty_one(uncertainty)
-        if uncertainty.ndim == 4:
-            if uncertainty.shape[1] != 1:
-                raise ValueError(f"Unexpected input shape={uncertainty.shape}, expecting [N,1,H,W].")
-            return [visualize_uncertainty_one(img, idx) for idx, img in enumerate(uncertainty)]
-        raise ValueError(f"Unexpected input shape={uncertainty.shape}: expecting 2, 3, or 4 dimensions.")
-    elif isinstance(uncertainty, list):
-        return [visualize_uncertainty_one(img, idx) for idx, img in enumerate(uncertainty)]
-    else:
-        raise ValueError(f"Unexpected input type: {type(uncertainty)}")
